+++ src/node-addon/auth-context.hh
@@ -1,5 +1,5 @@
-#ifndef WINSSO_AUTH_CONTEXT_HH
-#define WINSSO_AUTH_CONTEXT_HH
+#ifndef AUTH_CONTEXT_HH
+#define AUTH_CONTEXT_HH
 
 #define SECURITY_WIN32
 
@@ -10,43 +10,43 @@
 #include <Secext.h>
 #include <Security.h>
 
-namespace WinSso {
-  class AuthContext {
-    public:
-    char packageName[128];
-    std::string targetHostname;
-    CredHandle credHandle;
-    SECURITY_INTEGER lifeTime;
-    struct _SecHandle ctxHandle;
-    unsigned long ctxAttributes;
-    unsigned long maxTokenLength;
-    unsigned char* outToken;
-    unsigned long outTokenLength;
+class AuthContext {
+  public:
+  char packageName[128];
+  std::string targetHostname;
+  CredHandle credHandle;
+  SECURITY_INTEGER lifeTime;
+  struct _SecHandle ctxHandle;
+  unsigned long ctxAttributes;
+  unsigned long maxTokenLength;
+  unsigned char* outToken;
+  unsigned long outTokenLength;
 
-    private:
-    struct SecChannelBindingsCombined {
-      struct _SEC_CHANNEL_BINDINGS secChannelBindings;
-      unsigned char applicationData[128];
-    };
-    struct SecChannelBindingsCombined channelBindings;
-    unsigned long channelBindingsLength;
-
-    bool credHandleAllocated;
-    bool ctxHandleAllocated;
-
-    public:
-    AuthContext();
-    virtual ~AuthContext();
-    void Cleanup(Napi::Env* env);
-
-    bool Init(std::string* packageName, std::string* targetHost, Napi::Buffer<unsigned char>& applicationDataBuffer, Napi::Env* env);
-    bool InitContext(Napi::Env* env);
-    bool HandleResponse(Napi::Buffer<unsigned char>& inTokenBuffer, Napi::Env* env);
-
-    Napi::Buffer<unsigned char> OutToken(Napi::Env* env);
-
-    private:
-    void SetupChannelBindings(Napi::Buffer<unsigned char>& applicationDataBuffer);
+  private:
+  struct SecChannelBindingsCombined {
+    struct _SEC_CHANNEL_BINDINGS secChannelBindings;
+    unsigned char applicationData[128];
   };
-}
+  struct SecChannelBindingsCombined channelBindings;
+  unsigned long channelBindingsLength;
+
+  bool credHandleAllocated;
+  bool ctxHandleAllocated;
+
+  public:
+  AuthContext();
+  virtual ~AuthContext();
+  void Cleanup(Napi::Env* env);
+
+  bool Init(std::string* packageName, std::string* targetHost, Napi::Buffer<unsigned char>& applicationDataBuffer, Napi::Env* env);
+  bool InitContext(Napi::Env* env);
+  bool HandleResponse(Napi::Buffer<unsigned char>& inTokenBuffer, Napi::Env* env);
+
+  Napi::Buffer<unsigned char> OutToken(Napi::Env* env);
+
+  private:
+  void SetupChannelBindings(Napi::Buffer<unsigned char>& applicationDataBuffer);
+  std::string* TargetHostnameRef();
+};
+
 #endif