+++ test/win-sso.spec.ts
@@ -10,10 +10,10 @@ describe('WinSso', function() {
     }
   });
 
-  describe('getUserName', function () {
+  describe('getLogonUserName', function () {
     it('should return a username', function() {
       // Act
-      let result = WinSso.getUserName();
+      let result = WinSso.getLogonUserName();
 
       // Assert
       chai.expect(result.length).to.be.greaterThan(0);
@@ -22,9 +22,19 @@ describe('WinSso', function() {
   });
 
   describe('createAuthRequest', function () {
+    let winSso: WinSso;
+
+    beforeEach(function() {
+      winSso = new WinSso('NTLM', undefined, undefined);
+    });
+
+    afterEach(function() {
+      winSso.freeAuthContext();
+    });
+
     it('should return a token buffer', function() {
       // Act
-      let result = WinSso.createAuthRequest();
+      let result = winSso.createAuthRequest();
 
       // Assert
       chai.expect(result).to.be.instanceOf(Buffer);
@@ -32,7 +42,7 @@ describe('WinSso', function() {
 
     it('should not return an empty token buffer', function() {
       // Act
-      let result = WinSso.createAuthRequest();
+      let result = winSso.createAuthRequest();
 
       // Assert
       chai.expect(result.length).to.be.greaterThan(0);
@@ -40,7 +50,7 @@ describe('WinSso', function() {
 
     it('should provide a NTLM type 1 message', function() {
       // Act
-      let result = WinSso.createAuthRequest();
+      let result = winSso.createAuthRequest();
 
       // Assert
       let base64tokenHeader = result.slice(0,12).toString('base64');
@@ -50,9 +60,19 @@ describe('WinSso', function() {
   });
 
   describe('createAuthRequestHeader', function () {
+    let winSso: WinSso;
+
+    beforeEach(function() {
+      winSso = new WinSso('NTLM', undefined, undefined);
+    });
+
+    afterEach(function() {
+      winSso.freeAuthContext();
+    });
+
     it('should return a token header', function() {
       // Act
-      let result = WinSso.createAuthRequestHeader();
+      let result = winSso.createAuthRequestHeader();
 
       // Assert
       chai.expect(result.length).to.be.greaterThan(0);
@@ -60,7 +80,7 @@ describe('WinSso', function() {
 
     it('should prefix the token with \'NTLM \'', function() {
       // Act
-      let result = WinSso.createAuthRequestHeader();
+      let result = winSso.createAuthRequestHeader();
 
       // Assert
       chai.expect(result.indexOf('NTLM ')).to.equal(0);
@@ -68,8 +88,8 @@ describe('WinSso', function() {
 
     it('should provide a base64 encoded token from createAuthRequest', function() {
       // Act
-      let result = WinSso.createAuthRequestHeader();
-      let token = WinSso.createAuthRequest();
+      let result = winSso.createAuthRequestHeader();
+      let token = winSso.createAuthRequest();
 
       // Assert
       chai.expect(result.substring(5)).to.equal(token.toString('base64'));
@@ -79,10 +99,21 @@ describe('WinSso', function() {
   describe('createAuthResponse', function () {
     const type2MessageHeader = 'NTLM TlRMTVNTUAACAAAAFAAUADgAAAAFAIkCU3J2Tm9uY2UAAAAAAAAAAJYAlgBMAAAACgC6RwAAAA9VAFIAUwBBAC0ATQBJAE4ATwBSAAEAEABNAE8AUwBJAFMATABFAFkAAgAUAFUAUgBTAEEALQBNAEkATgBPAFIAAwAmAE0AbwBzAEkAcwBsAGUAeQAuAHUAcgBzAGEALgBtAGkAbgBvAHIABAAUAHUAcgBzAGEALgBtAGkAbgBvAHIABQAUAHUAcgBzAGEALgBtAGkAbgBvAHIABwAIAKUvIxkwMNUBAAAAAA==';
     const targetHost = 'MosIsley.ursa.minor';
+    let winSso: WinSso;
+
+    beforeEach(function() {
+      winSso = new WinSso('NTLM', targetHost, undefined);
+    });
+
+    afterEach(function() {
+      winSso.freeAuthContext();
+    });
 
     it('should return a token buffer', function() {
+      winSso.createAuthRequest();
+
       // Act
-      let result = WinSso.createAuthResponse(type2MessageHeader, targetHost, undefined);
+      let result = winSso.createAuthResponse(type2MessageHeader);
 
       // Assert
       chai.expect(result).to.be.instanceOf(Buffer);
@@ -90,21 +121,25 @@ describe('WinSso', function() {
     });
 
     it('should not return an empty token buffer', function() {
+      winSso.createAuthRequest();
+
       // Act
-      let result = WinSso.createAuthResponse(type2MessageHeader, targetHost, undefined);
+      let result = winSso.createAuthResponse(type2MessageHeader);
 
       // Assert
       chai.expect(result.length).to.be.greaterThan(0);
     });
 
     it('should provide a NTLM type 3 message', function() {
+      winSso.createAuthRequest();
+
       // Act
-      let result = WinSso.createAuthResponse(type2MessageHeader, targetHost, undefined);
+      let result = winSso.createAuthResponse(type2MessageHeader);
 
       // Assert
       let base64tokenHeader = result.slice(0,12).toString('base64');
-      let expectType1Header = Buffer.from("NTLMSSP\0\x03\x00\x00\x00").toString('base64');
-      chai.expect(base64tokenHeader).to.equal(expectType1Header);
+      let expectType3Header = Buffer.from("NTLMSSP\0\x03\x00\x00\x00").toString('base64');
+      chai.expect(base64tokenHeader).to.equal(expectType3Header);
     });
 
     it('should provide a NTLM type 3 message when passed a PeerCertificate', function() {
@@ -113,49 +148,111 @@ describe('WinSso', function() {
       let fakeCert: any = {
         fingerprint256: fingerprint
       };
+      let winSsoCert = new WinSso('NTLM', targetHost, fakeCert);
+      winSsoCert.createAuthRequest();
 
       // Act
-      let result = WinSso.createAuthResponse(type2MessageHeader, targetHost, fakeCert);
+      let result = winSso.createAuthResponse(type2MessageHeader);
+      winSsoCert.freeAuthContext();
 
       // Assert
       let base64tokenHeader = result.slice(0,12).toString('base64');
-      let expectType1Header = Buffer.from("NTLMSSP\0\x03\x00\x00\x00").toString('base64');
-      chai.expect(base64tokenHeader).to.equal(expectType1Header);
+      let expectType3Header = Buffer.from("NTLMSSP\0\x03\x00\x00\x00").toString('base64');
+      chai.expect(base64tokenHeader).to.equal(expectType3Header);
+    });
+
+    it('should provide a NTLM type 3 message with undefined targetHost', function() {
+      winSso.createAuthRequest();
+
+      // Act
+      let result = winSso.createAuthResponse(type2MessageHeader);
+
+      // Assert
+      let base64tokenHeader = result.slice(0,12).toString('base64');
+      let expectType3Header = Buffer.from("NTLMSSP\0\x03\x00\x00\x00").toString('base64');
+      chai.expect(base64tokenHeader).to.equal(expectType3Header);
+    });
+
+    // This test accepts two scenarios - success or throws a specific error. The background is that
+    // we don't know the client settings where the test is executed - it might allow NTLMv1
+    it('should provide a NTLM type 3 message from NTLM v1 challenge, if NTLMv1 is allowed by client', function() {
+      // Arrange
+      let ntlmV1_type2MessageHeader = 'NTLM TlRMTVNTUAACAAAAAAAAAAAoAAABggAAASNFZ4mrze8AAAAAAAAAAA==';
+      const expectNtlmV1error = 'Could not create NTLM type 3 message. Incoming type 2 message uses NTLMv1, it is likely that the client is prevented from sending such messages. Update target host to use NTLMv2 (recommended) or adjust LMCompatibilityLevel on the client (insecure)';
+      let winSso: WinSso;
+
+      beforeEach(function() {
+        winSso = new WinSso('NTLM', targetHost, undefined);
+      });
+
+      afterEach(function() {
+        winSso.freeAuthContext();
+      });
+
+      try {
+        // Act
+        winSso.createAuthRequest();
+        let result = winSso.createAuthResponse(ntlmV1_type2MessageHeader);
+
+        // Assert
+        let base64tokenHeader = result.slice(0,12).toString('base64');
+        let expectType3Header = Buffer.from("NTLMSSP\0\x03\x00\x00\x00").toString('base64');
+        chai.expect(base64tokenHeader).to.equal(expectType3Header);
+      } catch (err) {
+        // Assert
+        chai.expect(err.message).to.equal(expectNtlmV1error);
+      }
     });
 
     it('should throw if inToken is not a NTLM type 2 message', function() {
       // Arrange
       let dummyToken = 'NOT NTLM AT ALL';
+      winSso.createAuthRequest();
 
       // Act & Assert
-      chai.expect(() => WinSso.createAuthResponse(dummyToken, targetHost, undefined)).to.throw();
+      chai.expect(() => winSso.createAuthResponse(dummyToken)).to.throw();
     });
   });
 
   describe('createAuthResponseHeader', function () {
     const type2MessageHeader = 'NTLM TlRMTVNTUAACAAAAFAAUADgAAAAFAIkCU3J2Tm9uY2UAAAAAAAAAAJYAlgBMAAAACgC6RwAAAA9VAFIAUwBBAC0ATQBJAE4ATwBSAAEAEABNAE8AUwBJAFMATABFAFkAAgAUAFUAUgBTAEEALQBNAEkATgBPAFIAAwAmAE0AbwBzAEkAcwBsAGUAeQAuAHUAcgBzAGEALgBtAGkAbgBvAHIABAAUAHUAcgBzAGEALgBtAGkAbgBvAHIABQAUAHUAcgBzAGEALgBtAGkAbgBvAHIABwAIAKUvIxkwMNUBAAAAAA==';
     const targetHost = 'MosIsley.ursa.minor';
+    let winSso: WinSso;
+
+    beforeEach(function() {
+      winSso = new WinSso('NTLM', targetHost, undefined);
+    });
+
+    afterEach(function() {
+      winSso.freeAuthContext();
+    });
 
     it('should return a token header', function() {
+      winSso.createAuthRequest();
+
       // Act
-      let result = WinSso.createAuthResponseHeader(type2MessageHeader, targetHost, undefined);
+      let result = winSso.createAuthResponseHeader(type2MessageHeader);
 
       // Assert
       chai.expect(result.length).to.be.greaterThan(0);
     });
 
     it('should prefix the token with \'NTLM \'', function() {
+      winSso.createAuthRequest();
+
       // Act
-      let result = WinSso.createAuthResponseHeader(type2MessageHeader, targetHost, undefined);
+      let result = winSso.createAuthResponseHeader(type2MessageHeader);
 
       // Assert
       chai.expect(result.indexOf('NTLM ')).to.equal(0);
     });
 
     it('should provide a base64 encoded token from createAuthResponse', function() {
+      winSso.createAuthRequest();
+
       // Act
-      let result = WinSso.createAuthResponseHeader(type2MessageHeader, targetHost, undefined);
-      let token = WinSso.createAuthResponse(type2MessageHeader, targetHost, undefined);
+      let result = winSso.createAuthResponseHeader(type2MessageHeader);
+      let token = winSso.createAuthResponse(type2MessageHeader);
 
       // Assert
       // Since the token will contain unique challenge and timestamp values the two calls